minor things renaming writeShoppingListToCSV, eliminating parts of test and main that were not up to date, incorporate new BigDecimal status to AddAmountFromShopping.java test

+++ src/com/louisa/test/AddAmountFromShopping.java
@@ -1,6 +1,9 @@
 package com.louisa.test;
 import com.louisa.butlerpantry.Ingredient;
 import org.junit.jupiter.api.Test;
+
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 
@@ -9,15 +12,15 @@ public class AddAmountFromShopping {
 
     @Test
     public void testAddAmountFromShopping() {
-        Ingredient testIngredient = new Ingredient("Flour", "g", 1000.0);
-        testIngredient.addAmountFromShopping(500.0);
-        assertEquals(1500.0, testIngredient.getAmount());
+        Ingredient testIngredient = new Ingredient("Flour", "g", BigDecimal.valueOf(1000.0));
+        testIngredient.addAmountFromShopping(BigDecimal.valueOf(500.0));
+        assertEquals(BigDecimal.valueOf(1500), testIngredient.getAmount());
     }
 
     @Test
     public void testAddNegativeAmount() {
-        Ingredient testIngredient = new Ingredient("Flour", "g", 50.0);
-        testIngredient.addAmountFromShopping(-0.2);
+        Ingredient testIngredient = new Ingredient("Flour", "g", BigDecimal.valueOf(50));
+        testIngredient.addAmountFromShopping(BigDecimal.valueOf(-0.2));
         assertEquals(49.8, testIngredient.getAmount());
     }
 }
